+++ src/yafowil/widget/ace/example.py
@@ -136,7 +136,7 @@ DOC_ACE_DARKMODE = """
 Dark Mode
 ---------
 
-Customize syntax highlighting and appearance  for dark mode with one of the
+Customize syntax highlighting and appearance for dark mode with one of the
 available ACE themes.
 The current ACE theme corresponds to the theme set in Bootstrap5 HTML
 data-bs-theme attribute ('light' | 'dark').
@@ -186,10 +186,53 @@ def ace_darkmode():
     }
 
 
+DOC_ACE_DISPLAY = """
+Display Mode / Read-only Mode
+-----------------------------
+
+The widget attribute ``read_only`` can be used to render the widget in read-only mode.
+
+The widget's display renderer automatically sets the ``read_only`` attribute to ``True``.
+
+.. code-block:: python
+
+    value = '''
+    function foo(items) {
+        for (var i=0; i<=items.length; i++) {
+            alert(items[i]);
+        }
+    }
+    '''
+    ace = factory('#field:ace', value=value, mode='display', props={
+        'label': 'Light and Dark Theme',
+        'required': 'Code is required',
+        'theme': 'chrome',
+        'dark_theme': 'tomorrow_night',
+        'mode': 'javascript',
+        # 'read_only': True # manually render in read-only mode
+    })
+"""
+
+
+def ace_display():
+    part = factory(u'fieldset', name='yafowil.widget.ace.display')
+    part['ace'] = factory('#field:ace', value=ace_theme_value, mode='display', props={
+        'label': 'Display Mode',
+        'required': 'Code is required',
+        'theme': 'chrome',
+        'dark_theme': 'tomorrow_night',
+        'mode': 'javascript'
+    })
+    return {
+        'widget': part,
+        'doc': DOC_ACE_DISPLAY,
+        'title': 'Display Mode',
+    }
+
 
 def get_example():
     themes = []
     if factory.theme == 'bootstrap5':
         themes.append(ace_darkmode())
-    themes += [ace_python(), ace_js(), ace_themes()]
+    themes += [ace_python(), ace_js(), ace_themes(), ace_display()]
     return themes